% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.R
\name{DiffGini}
\alias{DiffGini}
\title{Inference on the difference of Gini Indices}
\usage{
DiffGini(x, y, basis, method = "optimal", CItype, tau = 0.05,
  B = 1000)
}
\arguments{
\item{x}{observed sample from \eqn{F_0}.}

\item{y}{observed sample from \eqn{F_1}.}

\item{basis}{pre-sepecified basis function \eqn{\boldsymbol{q}(x)} in the exponential term.
It should return a \eqn{d}-dimensional vector.
For example, \code{basis = function(x){c(x,log(x))}}.}

\item{method}{the method used to fit the DRM. The options include \code{"glm"}, \code{"multinom"} and \code{"optimal"}.
The default setting is \code{"optimal"}. See \code{help(DRM)} for `Details'.}

\item{CItype}{the method to be used for confidence interval construction. See `Details'.}

\item{tau}{the significance level. The defualt value is \eqn{0.05}.}

\item{B}{the number of bootstap samples. The defualt value is \eqn{1000}.}
}
\value{
The function returns a vector containing the following components:
\itemize{
\item \code{estimate}: the estimate of the difference \eqn{\mathcal{G}_0 - \mathcal{G}_1}.
\item \code{ASD}: the asymptotic standard deviation (ASD) of the estimator of the difference.
\item \code{lower bound}: the lower bound of the confidence interval for the difference.
\item \code{upper bound}: the upper bound of the confidence interval for the difference.
}
When \code{CItype = "None"}, it only returns the estimate and ASD of the difference.
}
\description{
Estimate the difference of Gini indices as well as construct the confidence interval for the different
under the density ratio model (DRM) based on the observed samples.
}
\details{
Let \eqn{\mathcal{G}_i} be the Gini index of population \eqn{i}.
The argument \code{CItype} refers to difference confidence intervals for the difference \eqn{\mathcal{G}_0 - \mathcal{G}_1}
under the DRMs in \insertCite{yuan2021Gini;textual}{ZeroDRM}.
\itemize{
\item  \code{"None"}: no confidence intervals for the difference is constructed;
\item  \code{"NA-DRM"}: the Wald-type confidence intervals for the difference based on the normal approximation;
\item  \code{"BT-DRM"}: the bootstrap-t confidence intervals for the difference ;
}
}
\examples{
data(Ilocos,package = "ineq")
x = Ilocos$income[Ilocos$urban == "urban"&Ilocos$pro == "Pangasinan"]
y = Ilocos$income[Ilocos$urban == "rural"&Ilocos$pro == "Pangasinan"]
basis = function(x){log(x)}
DiffGini(x,y,basis,CItype = "NA-DRM")

}
\references{
\insertRef{yuan2021Gini}{ZeroDRM}
}
