% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{MeanRatio}
\alias{MeanRatio}
\title{Inference on mean ratio}
\usage{
MeanRatio(x, y, basis, method = "optimal", CItype, tau = 0.05)
}
\arguments{
\item{x}{observed sample from \eqn{F_0}.}

\item{y}{observed sample from \eqn{F_1}.}

\item{basis}{pre-sepecified basis function \eqn{\boldsymbol{q}(x)} in the exponential term.
It should return a \eqn{d}-dimensional vector.
For example, \code{basis = function(x){c(x,log(x))}}.}

\item{method}{the method used to fit the DRM. The options include \code{"glm"}, \code{"multinom"} and \code{"optimal"}.
The default setting is \code{"optimal"}. See \code{help(DRM)} for `Details'.}

\item{CItype}{the method to be used for confidence interval construction. See `Details'.}

\item{tau}{the significance level. The defualt value is \eqn{0.05}.}
}
\value{
The function returns a vector containing the following components:
\itemize{
\item \code{estimate}: the estimate of the difference \eqn{\mathcal{G}_0 - \mathcal{G}_1}.
\item \code{ASD}: the asymptotic standard deviation (ASD) of the estimator of the difference.
\item \code{lower bound}: the lower bound of the confidence interval for the difference.
\item \code{upper bound}: the upper bound of the confidence interval for the difference.
}
When \code{CItype = "None"}, it only returns the estimate and ASD of the difference.
}
\description{
Estimate the mean ratio and construct the confidence interval based on the observed samples.
}
\details{
Let \eqn{\mu_i = \int_0^{\infty} x dF_i(x)} be the mean of populatio \eqn{i}.
Then we define the functional \eqn{\boldsymbol{\psi}} in \insertCite{yuan2020zeros;textual}{ZeroDRM} as
 \eqn{\boldsymbol{\psi} = (\mu_0,\mu_1)^\top}.

Further let the function \eqn{g(\cdot)} in  \insertCite{yuan2020zeros;textual}{ZeroDRM} as
\eqn{g(x_1,x_2) = x_2/x_1}. We then have \eqn{g(\boldsymbol{\psi}) = \mu_1/\mu_0} is the mean ratio of two populations.

The argument \code{CItype} refers to difference confidence intervals for the mean ratio \eqn{\mu_1/\mu_0}
under the DRMs in \insertCite{yuan2020zeros;textual}{ZeroDRM}.
\itemize{
\item  \code{"None"}: no confidence intervals for the difference is constructed;
\item  \code{"NA-DRM"}: the Wald-type confidence intervals based on the normal approximation of \eqn{\mu_1/\mu_0};
\item  \code{"log-DRM"}: the Wald-type confidence intervals based on the normal approximation of \eqn{\log(\mu_1/\mu_0)}.
}
}
\references{
\insertRef{yuan2020zeros}{ZeroDRM}
}
