% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{ZeroFunc}
\alias{ZeroFunc}
\title{Inference on Functionals}
\usage{
ZeroFunc(x, y, basis, method = "optimal", Ufun, Pnu, Ptheta)
}
\arguments{
\item{x}{observed sample from \eqn{F_0}.}

\item{y}{observed sample from \eqn{F_1}.}

\item{basis}{pre-sepecified basis function \eqn{\boldsymbol{q}(x)} in the exponential term.
It should return a \eqn{d}-dimensional vector.
For example, \code{basis = function(x){c(x,log(x))}}.}

\item{method}{the method used to fit the DRM. The options include \code{"glm"}, \code{"multinom"} and \code{"optimal"}.
The default setting is \code{"optimal"}. See \code{help(DRM)} for `Details'.}

\item{Ufun}{the funtion \eqn{\boldsymbol{u}(x;\boldsymbol{\nu},\boldsymbol{\theta})} to specify the \eqn{p}-demensional functional \eqn{\boldsymbol{\psi}}
with zero proportion parameter \eqn{\boldsymbol{\nu} = (\nu_0,\nu_1)^\top} and DRM parameter \eqn{\boldsymbol{\theta} = (\alpha,\boldsymbol{\beta})^\top}.
It should return a \eqn{p}-dimensional vector. See `Details`.}

\item{Pnu}{a function which specifies the first derivative of \eqn{\boldsymbol{u}(x;\boldsymbol{\nu},\boldsymbol{\theta})} with respect to \eqn{\boldsymbol{\nu}}.
It should return a \eqn{p\times 2} matrix. See `Details`.}

\item{Ptheta}{a function which specifies the first derivative of function \eqn{\boldsymbol{u}(x;\boldsymbol{\nu},\boldsymbol{\theta})} with respect to \eqn{\boldsymbol{\theta}}.
It should return a \eqn{p\times (d+1)} matrix, where \eqn{d} is the dimensional of \eqn{\boldsymbol{\theta}}.
See `Details`.}
}
\value{
The function returns a list containing the following components:
\itemize{
\item \code{estimate}: the estimates of functional \eqn{\boldsymbol{\psi}}.
\item \code{AVar}: the estimate of asymptotic variance-covariance matrix \eqn{\Gamma}.
The explicit form of \eqn{\Gamma} van be found in \insertCite{yuan2020zeros;textual}{ZeroDRM}.
}
}
\description{
Estimate the functionals as well as construct the confidence interval for functionals based on the observed samples.
}
\details{
\insertCite{yuan2020zeros;textual}{ZeroDRM} considered a class of general functionals \eqn{\boldsymbol{\psi}} of dimension \eqn{p}, defined as
\deqn{\boldsymbol{\psi}= \int_0^\infty \boldsymbol{u}(x;\boldsymbol{\nu},\boldsymbol{\theta}) dG_0(x),}
where \eqn{\nu=(\nu_0,\nu_1)^\top}, \eqn{\boldsymbol{\theta} = (\alpha,\boldsymbol{\beta})^\top},
and  \eqn{\boldsymbol{u}(x;\boldsymbol{\nu},\boldsymbol{\theta})=\left(u_1(x;\boldsymbol{\nu},\boldsymbol{\theta}),\ldots,u_p(x;\boldsymbol{\nu},\boldsymbol{\theta})\right)^\top} is a given \eqn{(p\times 1)}-dimensional function.

\code{Pnu} refers to the the first derivative of \eqn{\boldsymbol{u}(x;\boldsymbol{\nu},\boldsymbol{\theta})} with respect to \eqn{\boldsymbol{\nu}}:
\deqn{\frac{\partial \boldsymbol{u}(x;\boldsymbol{\nu},\boldsymbol{\theta})}{\partial \boldsymbol{\nu}}
= \left(\begin{array}{cc}
\frac{\partial u_1(x;\boldsymbol{\nu},\boldsymbol{\theta})}{\partial \nu_0}&\frac{\partial u_1(x;\boldsymbol{\nu},\boldsymbol{\theta})}{\partial \nu_1}\\
\vdots&\vdots\\
\frac{\partial u_p(x;\boldsymbol{\nu},\boldsymbol{\theta})}{\partial \nu_0}&\frac{\partial u_p(x;\boldsymbol{\nu},\boldsymbol{\theta})}{\partial \nu_1}
\end{array}\right).}

\code{Ptheta} refers to the the first derivative of \eqn{\boldsymbol{u}(x;\boldsymbol{\nu},\boldsymbol{\theta})} with respect to \eqn{\boldsymbol{\theta}}:
\deqn{\frac{\partial \boldsymbol{u}(x;\boldsymbol{\nu},\boldsymbol{\theta})}{\partial \boldsymbol{\theta}}
= \left(\begin{array}{cc}
\frac{\partial u_1(x;\boldsymbol{\nu},\boldsymbol{\theta})}{\partial \alpha}&\frac{\partial u_1(x;\boldsymbol{\nu},\boldsymbol{\theta})}{\partial \boldsymbol{\beta}}\\
\vdots&\vdots\\
\frac{\partial u_p(x;\boldsymbol{\nu},\boldsymbol{\theta})}{\partial \alpha}&\frac{\partial u_p(x;\boldsymbol{\nu},\boldsymbol{\theta})}{\partial \boldsymbol{\beta}}
\end{array}\right).}
}
\references{
\insertRef{yuan2020zeros}{ZeroDRM}
}
