% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.R
\name{ZeroGini}
\alias{ZeroGini}
\title{Inference on Gini Indices}
\usage{
ZeroGini(x, y, basis, method = "optimal", CItype, tau = 0.05,
  B = 1000)
}
\arguments{
\item{x}{observed sample from \eqn{F_0}.}

\item{y}{observed sample from \eqn{F_1}.}

\item{basis}{pre-sepecified basis function \eqn{\boldsymbol{q}(x)} in the exponential term.
It should return a \eqn{d}-dimensional vector.
For example, \code{basis = function(x){c(x,log(x))}}.}

\item{method}{the method used to fit the DRM. The options include \code{"glm"}, \code{"multinom"} and \code{"optimal"}.
The default setting is \code{"optimal"}. See \code{help(DRM)} for `Details'.}

\item{CItype}{the method to be used for confidence interval construction. See `Details'.}

\item{tau}{the significance level. The defualt value is \eqn{0.05}.}

\item{B}{the number of bootstap samples. The defualt value is \eqn{1000}.}
}
\value{
The function returns a list containing the following components:
\itemize{
\item \code{Gini0}: the estimate and the confidence interval (lower bound and uppper bound) of \eqn{\mathcal{G}_0}.
When \code{CItype = "None"}, it only returns the estimate of \eqn{\mathcal{G}_0}.
\item \code{Gini1}: the estimate and the confidence interval (lower bound and uppper bound) of \eqn{\mathcal{G}_1}.
When \code{CItype = "None"}, it only returns the estimate of \eqn{\mathcal{G}_1}.
\item \code{var}: the estimate of asymptotic variance-covariance matrix \eqn{\Sigma}.
 The explicit form of \eqn{\Sigma} van be found in \insertCite{yuan2021Gini;textual}{ZeroDRM}.
}
}
\description{
Estimate the Gini indices as well as construct the confidence intervals for the Gini indices
under the density ratio model (DRM) based on the observed samples.
}
\details{
The Gini index of the semicontinuous population \eqn{i} can be expessed by
\deqn{\mathcal{G}_i=(2\nu_i-1)+(1-\nu_i)\frac{\int_0^{\infty} \{2xG_i(x)\}dG_i(x)}{ \int_0^{\infty} x dG_i(x)}.}
\insertCite{yuan2021Gini;textual}{ZeroDRM} estimated the Gini indices \eqn{\mathcal{G}_0} and \eqn{\mathcal{G}_1} based on the above expression.
When there is no excess of zero values in the data, the proposed method is also applicable by setting \eqn{\nu_i = 0}.

The argument \code{CItype} refers to difference confidence intervals of Gini indices
under the DRMs in \insertCite{yuan2021Gini;textual}{ZeroDRM}.
\itemize{
\item  \code{"None"}: no confidence intervals for Gini indices is constructed;
\item  \code{"NA-DRM"}: the Wald-type confidence intervals based on the normal approximation;
\item  \code{"NL-DRM"}: the Wald-type confidence intervals using logit transformation;
\item  \code{"BT-DRM"}: the bootstrap-t confidence intervals;
\item  \code{"BL-DRM"}: the bootstrap-t confidence intervals using logit transformation.
}
}
\examples{
data(Ilocos,package = "ineq")
x = Ilocos$income[Ilocos$urban == "urban"&Ilocos$pro == "Pangasinan"]
y = Ilocos$income[Ilocos$urban == "rural"&Ilocos$pro == "Pangasinan"]
basis = function(x){log(x)}
ZeroGini(x,y,basis,CItype = "NA-DRM")

}
\references{
\insertRef{yuan2021Gini}{ZeroDRM}
}
